+++ app/page.tsx
@@ -1,5 +1,4 @@
 import Landing from "@/components/Landing/Landing";
-import Image from "next/image";
 
 export default function Home() {
   return (
