+++ package.json
@@ -8,17 +8,13 @@
     "start": "next start",
     "lint": "next lint"
   },
-  "engines": {
-    "node": "20.x"
-  },
   "dependencies": {
     "@fortawesome/free-brands-svg-icons": "^6.6.0",
     "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "autoprefixer": "^10.4.20",
     "fortawesome": "^0.0.1-security",
-    "framer-motion": "^11.11.9",
-    "next": "^14.2.15",
+    "next": "^14.2.6",
     "next-images": "^1.8.5",
     "react": "^18",
     "react-dom": "^18",
@@ -28,9 +24,12 @@
     "util": "^0.12.5"
   },
   "devDependencies": {
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.14"
+    "tailwindcss": "^3.4.1",
+    "eslint-config-next": "14.2.16"
   }
 }
